replace individual add/remove commands with toggles

+++ cogs/AutoRole.py
@@ -1,7 +1,7 @@
 from math import ceil
 
 import discord
-from discord import app_commands
+from discord import app_commands, Interaction
 from discord.ext import commands
 
 
@@ -16,12 +16,12 @@ class AutoRoleList(discord.ui.View):
         await self.response.edit(view = None)
 
     @discord.ui.button(label = 'First page', style = discord.ButtonStyle.red)
-    async def first(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def first(self, interaction: Interaction, button: discord.ui.Button):
         await interaction.response.edit_message(embed = self.embedlist[0])
         self.page = 0
 
     @discord.ui.button(label = 'Previous page', style = discord.ButtonStyle.red)
-    async def previous(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def previous(self, interaction: Interaction, button: discord.ui.Button):
         if self.page >= 1:
             self.page -= 1
             await interaction.response.edit_message(embed = self.embedlist[self.page])
@@ -30,19 +30,19 @@ class AutoRoleList(discord.ui.View):
             await interaction.response.edit_message(embed = self.embedlist[self.page])
 
     @discord.ui.button(label = 'Stop', style = discord.ButtonStyle.grey)
-    async def exit(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def exit(self, interaction: Interaction, button: discord.ui.Button):
         await interaction.response.edit_message(embed = None)
         await self.on_timeout()
 
     @discord.ui.button(label = 'Next Page', style = discord.ButtonStyle.green)
-    async def next(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def next(self, interaction: Interaction, button: discord.ui.Button):
         self.page += 1
         if self.page > len(self.embedlist) - 1:
             self.page = 0
         await interaction.response.edit_message(embed = self.embedlist[self.page])
 
     @discord.ui.button(label = 'Last Page', style = discord.ButtonStyle.green)
-    async def last(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def last(self, interaction: Interaction, button: discord.ui.Button):
         self.page = len(self.embedlist) - 1
         await interaction.response.edit_message(embed = self.embedlist[self.page])
 
@@ -72,19 +72,18 @@ class AutoRole(commands.GroupCog, name = "autorole"):
     async def on_ready(self):
         print("AutoRole cog online")
         await self.bot.execute("CREATE TABLE IF NOT EXISTS autorole(guild BIGINT, role BIGINT UNIQUE )")
-
-    @app_commands.command(name = "add")
-    async def AutoRoleAdd(self, interaction: discord.Interaction, role: discord.Role):
-        await self.bot.execute("INSERT INTO autorole(guild, role) VALUES($1,$2) ON CONFLICT (role) DO NOTHING", interaction.guild.id, role.id)
-        await interaction.response.send_message(content = f"{role.name} successfully added to the AutoRole")
-
-    @app_commands.command(name = "remove")
-    async def AutoRoleRemove(self, interaction: discord.Interaction, role: discord.Role):
-        await self.bot.execute("DELETE FROM autorole WHERE guild=$1 AND role=$2", interaction.guild.id, role.id)
-        await interaction.response.send_message(content = f"{role.name} successfully removed to the AutoRole")
+        
+    @app_commands.command(name = "setup", description = "add or remove role from autorole")
+    async def AutoRole(self, interaction: Interaction, role: discord.Role):
+        if not await self.bot.fetchval("SELECT EXISTS(SELECT 1 FROM autorole WHERE guild = $1 AND role = $2)", interaction.guild_id, role.id):
+            await self.bot.execute("INSERT INTO autorole(guild, role) VALUES($1, $2) ON CONFLICT (role) DO NOTHING", interaction.guild.id, role.id)
+            await interaction.response.send_message(f"{role.name} successfully added to the AutoRole")
+        else:
+            await self.bot.execute("DELETE FROM autorole WHERE guild = $1 AND role = $2", interaction.guild.id, role.id)
+            await interaction.response.send_message(content = f"{role.name} successfully removed to the AutoRole")
 
     @app_commands.command(name = "list")
-    async def AutoRoleList(self, interaction: discord.Interaction):
+    async def AutoRoleList(self, interaction: Interaction):
         view = AutoRoleList(timeout = 60)
         roles = await self.bot.fetch("SELECT role FROM autorole WHERE guild=$1", interaction.guild.id)
         for role in roles:
@@ -102,7 +101,7 @@ class AutoRole(commands.GroupCog, name = "autorole"):
             role = member.guild.get_role(role["role"])
             try:
                 await member.add_roles(role)
-            except:
+            except Exception:
                 pass
